add PRIVMSG command

+++ srcs/commands.cpp
@@ -2,7 +2,7 @@
 
 
 Channel *findChan(std::string name, std::deque<Channel*> channels, bool &created){
-    for(int i = 0; i < channels.size(); i++){
+    for(size_t i = 0; i < channels.size(); i++){
         if(channels[i]->getName() == name)
             return channels[i];
     }
